+++ empymod/scripts/tmtemod.py
@@ -331,7 +331,7 @@ can use
 
 
 """
-# Copyright 2016-2021 The empymod Developers.
+# Copyright 2016-2021 The EMSiG community.
 #
 # This file is part of empymod.
 #